+++ NutriTEc-Backend/NutriTEc-Backend/Dtos/MeasurementDto.cs
@@ -2,19 +2,19 @@
 {
     public class MeasurementDto
     {
-        public int Height { get; set; }
+        public double Height { get; set; }
 
-        public int Fatpercentage { get; set; }
+        public double Fatpercentage { get; set; }
 
-        public int Musclepercentage { get; set; }
+        public double Musclepercentage { get; set; }
 
-        public int Weight { get; set; }
+        public double Weight { get; set; }
 
-        public int Waist { get; set; }
+        public double Waist { get; set; }
 
-        public int Neck { get; set; }
+        public double Neck { get; set; }
 
-        public int Hips { get; set; }
+        public double Hips { get; set; }
 
         public DateTime Revisiondate { get; set; }
     }
